2023년 8월 23일

+++ src/main/java/com/example/repository/GR/GrDopeRepository.java
@@ -1,6 +1,6 @@
 package com.example.repository.GR;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import javax.transaction.Transactional;
 
@@ -10,13 +10,13 @@ import org.springframework.stereotype.Repository;
 import com.example.entity.Dope;
 
 @Repository
-public interface GrDopeRepository extends JpaRepository<Dope, BigDecimal> {
+public interface GrDopeRepository extends JpaRepository<Dope, BigInteger> {
     
     // 삭제
     @Transactional
-    void deleteByEmailAndPost_no(String email, BigDecimal postno);
+    void deleteByEmailAndPost_no(String email, BigInteger postno);
 
-    Dope findByEmailAndPost_no(String email, BigDecimal postno);
+    Dope findByEmailAndPost_no(String email, BigInteger postno);
 
-    int countByEmailAndPost_no(String email, BigDecimal postno);
+    int countByEmailAndPost_no(String email, BigInteger postno);
 }
